Resolvendo problema de arredondamento indesejado na listagem de chopes

+++ backend/src/main/java/br/com/chadschoperia/service/dto/ViewBeerDto.java
@@ -13,7 +13,7 @@ public class ViewBeerDto implements Serializable {
 
 	private String name;
 
-	private Long stock;
+	private Double stock;
 
 	private String rfid;
